+++ appEditors/AppExcEditor.py
@@ -4195,6 +4195,7 @@ class AppExcEditorUI:
 
         # Editor
         self.exit_editor_button = QtWidgets.QPushButton(_('Exit Editor'))
+        self.exit_editor_button.setIcon(QtGui.QIcon(self.app.resource_location + '/power16.png'))
         self.exit_editor_button.setToolTip(
             _("Exit from Editor.")
         )