+++ appTools/ToolDrilling.py
@@ -1826,6 +1826,9 @@ class ToolDrilling(AppTool, Excellon):
             with self.app.proc_container.new(_("Generating CNC Code")):
                 a_obj.app_obj.new_object("cncjob", job_name, job_init)
 
+            # Switch notebook to Selected page
+            self.app.ui.notebook.setCurrentWidget(self.app.ui.selected_tab)
+
         # Create promise for the new name.
         self.app.collection.promise(job_name)